+++ lib/globals.dart
@@ -21,7 +21,7 @@ final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
 const String progVersion = '0.2.260813';
-const int buildNumber = 89;
+const int buildNumber = 90;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';